Move pkg/helm into pkg/controllers/chart
(cherry picked from commit da069715c14b245c8d2242073829bd0ed99a66f3)
