+++ pkg/controllers/chart/chart.go
@@ -1,4 +1,4 @@
-package helm
+package chart
 
 import (
 	"context"
