+++ pkg/controllers/chart/chart_test.go
@@ -1,4 +1,4 @@
-package helm
+package chart
 
 import (
 	"strings"
